pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "cutemarkets"
dynamic = ["version"]
description = "Official Python client for the CuteMarkets options market-data API"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "CuteMarkets" }]
keywords = [
  "cutemarkets",
  "options",
  "finance",
  "market-data",
  "trading",
  "quant",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Financial and Insurance Industry",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Office/Business :: Financial :: Investment",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "typing-extensions>=4.7; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "ruff>=0.5",
  "mypy>=1.10",
  "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://cutemarkets.com"
Documentation = "https://cutemarkets.com/docs"
Source = "https://github.com/cutemarkets/cutemarkets-python"

[tool.hatch.version]
path = "src/cutemarkets/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/cutemarkets"]

[tool.hatch.build.targets.sdist]
include = [
  "src/cutemarkets",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
